WIP : Refacto previous code
