+++ services/oauth_credentials/authentification.py
@@ -1,4 +1,3 @@
-import os
 import json
 from google.oauth2.credentials import Credentials
 from google.auth.transport.requests import Request
